# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tickets Management OpenAPI Documentation",
        "description": "Laravel Swagger OpenAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get list of all booked tickets",
                "operationId": "0082053c3590628e674a11ec0d1466e6",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a new ticket",
                "operationId": "93882881cca9046d8c6ddbbab7309b4b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "flight_id",
                                    "passenger_name",
                                    "passport"
                                ],
                                "properties": {
                                    "flight_id": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "passenger_name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "passport": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "seat": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/tickets/reset": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Reset tickets to default demo data",
                "operationId": "7ee68f5424f6d3c88aa88cedc888535a",
                "responses": {
                    "200": {
                        "description": "Table reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Table reset successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tickets/{id}": {
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Update seat for ticket",
                "operationId": "1010dcd7a1f352766df66431f1df4745",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "seat": {
                                        "type": "string",
                                        "example": "12A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seat updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete a ticket",
                "operationId": "c37b65ae6ff3ffa8c9e0c14ea4d6a9f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ticket deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Ticket": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "seat": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "canceled"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tickets",
            "description": "Tickets"
        }
    ]
}